Seperate shell command "calibrate" to be independent
