+++ program/shell/calibrate.h
@@ -1,0 +1,6 @@
+#ifndef __SHELL_CALIBRATE_H
+#define __SHELL_CALIBEATE_H
+
+void shell_calibrate(char parameter[][MAX_CMD_LEN], int par_cnt);
+
+#endif
